refactor: mv AttachConfig to mgr
